requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==3.2.12
django-haystack==3.1.1
importlib-metadata==4.11.1
Markdown==3.3.6
mysqlclient==2.1.0
Pygments==2.11.2
pytz==2021.3
sqlparse==0.4.2
thulac==0.2.1
tzdata==2021.5
Whoosh==2.7.4
zipp==3.7.0
